requirements.txt
crewai[tools]>=0.108.0
litellm>=1.0.0
python-dotenv>=1.1.0
requests>=2.31.0
pydantic>=2.10.0
mcp>=1.0.0
watchdog>=3.0.0
